% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty_drop_citeseq_mtx.R
\docType{data}
\name{empty_drop_citeseq_mtx}
\alias{empty_drop_citeseq_mtx}
\title{sample CITEseq protein data 87 protein by 8005 empty droplets}
\format{A \code{matrix} 87 protein by 8005 empty droplets confirmed by cell hashing and cross referencing sample genotypes.
\describe {
empty_drop_citeseq_mtx is a matrix of empty droplets from a CITEseq experiment using cell hashing antibodies called by Seurat's HTODemux function as negative cells, called by Demuxlet via referencing patient genotype data as ambiguous drops, and with less than 80 unique genes detected. This is used for robust estimation of the background distribution of each protein.
}}
\usage{
empty_drop_citeseq_mtx
}
\description{
A \code{matrix} of cells by antibodies Raw CITEseq data used for example scripts of the dsb package. raw data processed with CITE-seq-count https://hoohm.github.io/CITE-seq-Count/
}
\keyword{datasets}
