% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cells_citeseq_mtx.r
\docType{data}
\name{cells_citeseq_mtx}
\alias{cells_citeseq_mtx}
\title{sample CITEseq protein data 87 protein by 2872 cells}
\format{A matrix 87 protein by 2872 cells
\describe {
\item{cells_citeseq_mtx}{is an R matrix of cells as columns and 87 proteins as rows. It is a random even distribution of cells manually annotated from protein clustering of maximum of 100 cells per cluster on 30 clusters. Full celltype data annotations and more information is available in the referenced publication.}
}}
\usage{
cells_citeseq_mtx
}
\description{
A matrix of cells by antibodies Raw CITEseq data used for example scripts of the dsb package. raw data processed with CITE-seq-count https://hoohm.github.io/CITE-seq-Count/
}
\keyword{datasets}
