% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsb.r
\name{DSBNormalizeProtein}
\alias{DSBNormalizeProtein}
\title{the DSB normalization function}
\usage{
DSBNormalizeProtein(
  cell_protein_matrix,
  empty_drop_matrix,
  define.pseudocount = FALSE,
  pseudocount.use,
  denoise.counts = TRUE,
  use.isotype.control = TRUE,
  isotype.control.name.vec = NULL
)
}
\arguments{
\item{cell_protein_matrix}{the data to be normalized cells = columns, proteins = rows}

\item{empty_drop_matrix}{the empty droplets used for background correction}

\item{define.pseudocount}{TRUE FALSE : false by default users can supply a pseudocount besides the default 10 which is optimal for CITEseq data.}

\item{pseudocount.use}{the pseudocount to use if overriding the default pseudocount by setting efine.pseudocount to TRUE}

\item{denoise.counts}{account for a per cell background correction factor by getting the mean of negative proteins in that cell + or -  isotype controls}

\item{use.isotype.control}{should the denoising covariate include isotype controls - recommended but false by default}

\item{isotype.control.name.vec}{a vector of the names of the isotype control proteins exactly as in the data to be normalized}
}
\value{
a normalized matrix that can be added to any S4 object e.g. Seurat or SingleCellExperiment Object
}
\description{
the DSB normalization function
}
